+++ packages/f-checkout/src/components/Checkout.vue
@@ -10,7 +10,6 @@
             :heading="$t('errorMessages.errorHeading')">
             {{ genericErrorMessage }}
         </alert>
-
         <card
             :card-heading="title"
             is-rounded
@@ -24,7 +23,7 @@
                 :class="$style['c-checkout-form']"
                 @submit.prevent="onFormSubmit">
                 <form-field
-                    v-model="mobileNumber"
+                    v-model="customer.mobileNumber"
                     name="mobile-number"
                     data-test-id="input-mobile-number"
                     :label-text="$t('labels.mobileNumber')"
@@ -40,14 +39,9 @@
 
                 <address-block
                     v-if="isCheckoutMethodDelivery"
-                    v-model="address"
-                    :address="address"
                     data-test-id="address-block" />
 
-                <form-selector
-                    :order-method="serviceType"
-                    :times="times"
-                    data-test-id="selector" />
+                <form-selector data-test-id="selector" />
 
                 <user-note data-test-id="user-note" />
 
@@ -93,14 +87,16 @@ import '@justeat/f-error-message/dist/f-error-message.css';
 import FormField from '@justeat/f-form-field';
 import '@justeat/f-form-field/dist/f-form-field.css';
 
+import { mapState, mapActions } from 'vuex';
 import AddressBlock from './Address.vue';
 import FormSelector from './Selector.vue';
 import UserNote from './UserNote.vue';
 
 import { CHECKOUT_METHOD_DELIVERY, TENANT_MAP } from '../constants';
 import tenantConfigs from '../tenants';
-import CheckoutServiceApi from '../services/CheckoutServiceApi';
 import EventNames from '../event-names';
+
+import checkoutModule from '../store/checkout.module';
 
 export default {
     name: 'VueCheckout',
@@ -140,23 +136,8 @@ export default {
     data () {
         return {
             tenantConfigs,
-            firstName: '',
-            mobileNumber: null,
-            address: {
-                line1: null,
-                line2: null,
-                city: null,
-                postcode: null
-            },
             genericErrorMessage: null,
-            shouldDisableCheckoutButton: false,
-            checkoutId: '',
-            isFulfillable: true,
-            times: [],
-            messages: [],
-            notes: [],
-            notices: [],
-            serviceType: this.checkoutMethod
+            shouldDisableCheckoutButton: false
         };
     },
 
@@ -169,15 +150,26 @@ export default {
 
         Object.defineProperty($v, 'addressValidations', {
             enumerable: true,
-            get: () => this.$v.address
+            get: () => this.$v.fulfillment.address
         });
 
         return { $v };
     },
 
     computed: {
+        ...mapState('checkout', [
+            'id',
+            'serviceType',
+            'customer',
+            'fulfillment',
+            'notes',
+            'isFulfillable',
+            'notices',
+            'messages'
+        ]),
+
         name () {
-            return (this.firstName.charAt(0).toUpperCase() + this.firstName.slice(1));
+            return (this.customer.firstName.charAt(0).toUpperCase() + this.customer.firstName.slice(1));
         },
 
         title () {
@@ -191,8 +183,8 @@ export default {
             * The $dirty boolean changes to true when the user has focused/lost
             * focus on the input field.
             */
-            const isMobileNumberValid = this.$v.mobileNumber.required && this.$v.mobileNumber.numeric && this.$v.mobileNumber.minLength;
-            return !this.$v.mobileNumber.$dirty || isMobileNumberValid;
+            const isMobileNumberValid = this.$v.customer.mobileNumber.required && this.$v.customer.mobileNumber.numeric && this.$v.customer.mobileNumber.minLength;
+            return !this.$v.customer.mobileNumber.$dirty || isMobileNumberValid;
         },
 
         isCheckoutMethodDelivery () {
@@ -206,40 +198,78 @@ export default {
 
     watch: {
         async checkoutUrl () {
-            await this.getCheckout();
+            await this.loadCheckout();
         }
     },
 
     async mounted () {
-        await this.getCheckout();
+        await this.loadCheckout();
     },
 
+    created () {
+        if (!this.$store.hasModule('checkout')) {
+            this.$store.registerModule('checkout', checkoutModule);
+        }
+    },
+
+    /*
+        TODO: in the future, we should actually try to deregister modules.
+        However, right now, given we might have several instances of the same component, we don't want to remove the module for all of them.
+        We tried generating a dynamic module name (so we could add and remove a module per component),
+        but we couldn't get it to work. It needs more investigation when this is really needed, not now.
+    */
+    // beforeDestroy () {
+    // this.$store.unregisterModule('checkout');
+    // },
+
     methods: {
-        async callCheckoutServiceApi () {
-            /*
-            * Add user data to `checkoutData` and submit to Checkout endpoint
-            * If this is successful emit `CheckoutSuccess` event
-            */
-            const checkoutData = {
-                mobileNumber: this.mobileNumber
-            };
+        ...mapActions('checkout', [
+            'getCheckout',
+            'postCheckout'
+        ]),
 
-            if (this.isCheckoutMethodDelivery) {
-                checkoutData.address = this.address;
+        /**
+         * Submit the checkout details while emitting events to communicate its success or failure.
+         *
+         */
+        async submitCheckout () {
+            try {
+                const checkoutData = {
+                    mobileNumber: this.customer.mobileNumber
+                };
+
+                if (this.isCheckoutMethodDelivery) {
+                    checkoutData.address = this.fulfillment.address;
+                }
+
+                await this.postCheckout({
+                    url: 'myPostUrl',
+                    tenant: this.tenant,
+                    data: checkoutData,
+                    timeout: this.checkoutTimeout
+                });
+
+                this.$emit(EventNames.CheckoutSuccess);
+            } catch (thrownErrors) {
+                this.$emit(EventNames.CheckoutFailure, thrownErrors);
             }
-
-            await CheckoutServiceApi.submitCheckout(this.checkoutUrl, this.tenant, checkoutData, this.checkoutTimeout);
-
-            this.$emit(EventNames.CheckoutSuccess);
         },
 
-        async getCheckout () {
+        /**
+         * Load the checkout details while emitting events to communicate its success or failure.
+         *
+         */
+        async loadCheckout () {
             try {
-                const result = await CheckoutServiceApi.getCheckout(this.checkoutUrl, this.tenant, this.getCheckoutTimeout);
-                this.$emit(EventNames.CheckoutGetSuccess);
-                this.mapResponse(result.data);
+                await this.getCheckout({
+                    url: this.checkoutUrl,
+                    tenant: this.tenant,
+                    timeout: this.getCheckoutTimeout
+                });
+
+                this.$emit(EventNames.CheckoutGetSuccess); // TODO: Check these emitted events.
             } catch (thrownErrors) {
-                this.$emit(EventNames.CheckoutGetFailure, thrownErrors);
+                this.$emit(EventNames.CheckoutGetFailure, thrownErrors); // TODO: Check these emitted events.
             }
         },
 
@@ -268,7 +298,7 @@ export default {
         async onFormSubmit () {
             /*
             * Check for is valid - no inline messages
-            * If form is valid try to call `CheckoutServiceApi`
+            * If form is valid try to call `submitCheckout`
             * Catch and handle any errors
             */
             if (!this.isFormValid()) {
@@ -280,7 +310,7 @@ export default {
             this.shouldDisableCheckoutButton = true;
 
             try {
-                await this.callCheckoutServiceApi();
+                await this.submitCheckout();
             } catch (error) {
                 this.handleErrorState(error);
             } finally {
@@ -294,51 +324,33 @@ export default {
             */
             this.$v.$touch();
             return !this.$v.$invalid;
-        },
-
-        mapResponse (data) {
-            this.checkoutId = data.id;
-            this.isFulfillable = data.isFulfillable;
-            if (data.customer) {
-                this.firstName = data.customer.firstName;
-                this.mobileNumber = data.customer.phoneNumber;
-            }
-
-            this.times = data.fulfillment.times;
-            if (data.fulfillment.address) {
-                this.address.line1 = data.fulfillment.address.lines[0];
-                this.address.line2 = data.fulfillment.address.lines[1];
-                this.address.city = data.fulfillment.address.lines[3];
-                this.address.postcode = data.fulfillment.address.postalCode;
-            }
-
-            this.messages = data.messages;
-            this.notes = data.notes;
-            this.notices = data.notices;
-            this.serviceType = data.serviceType;
         }
     },
 
     validations () {
         const deliveryDetails = {
-            mobileNumber: {
-                required,
-                numeric,
-                minLength: minLength(10)
+            customer: {
+                mobileNumber: {
+                    required,
+                    numeric,
+                    minLength: minLength(10)
+                }
             }
         };
 
         if (this.isCheckoutMethodDelivery) {
-            deliveryDetails.address = {
-                line1: {
-                    required
-                },
-                city: {
-                    required
-                },
-                postcode: {
-                    required,
-                    isValidPostcode: validations.isValidPostcode
+            deliveryDetails.fulfillment = {
+                address: {
+                    line1: {
+                        required
+                    },
+                    city: {
+                        required
+                    },
+                    postcode: {
+                        required,
+                        isValidPostcode: validations.isValidPostcode
+                    }
                 }
             };
         }