+++ packages/f-checkout/src/demo/demo.js
@@ -1,4 +1,5 @@
 import Vue from 'vue';
+import Vuex from 'vuex';
 import { VueI18n } from '@justeat/f-globalisation';
 import VueCheckout from '../components/Checkout.vue';
 import { ENGLISH_LOCALE } from '../../../storybook/constants/globalisation';
@@ -9,6 +10,7 @@ CheckoutMock.setupDelivery('/checkout-delivery.json');
 Vue.config.productionTip = false;
 
 Vue.use(VueI18n);
+Vue.use(Vuex);
 
 const i18n = new VueI18n({
     locale: ENGLISH_LOCALE,
@@ -19,6 +21,7 @@ const i18n = new VueI18n({
 /* eslint-disable no-new */
 new Vue({
     i18n,
+    store: new Vuex.Store({}),
     render: h => h(VueCheckout, {
         props: {
             checkoutUrl: '/checkout-delivery.json'