+++ packages/f-searchbox/CHANGELOG.md
@@ -3,6 +3,31 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+Latest (to be added to next release)
+------------------------------
+*November 26, 2020*
+
+### Added
+- css/scss to `moduleNameMapper` in `jest.config.js`
+
+
+v4.0.0-beta.4
+------------------------------
+*November 25, 2020*
+
+### Added
+
+- demo.js & index.html to serve up local instance of the component.
+
+### Removed
+
+- Demo.vue in favour of demo.js & index.html
+
+### Fixed
+
+- Vuex store mappings in form component, now using `mapState` & `mapActions`.
+
+
 v4.0.0-beta.3
 ------------------------------
 *November 20, 2020*